refactor(client): extract carjack code into its own file
