+++ fxmanifest.lua
@@ -15,7 +15,8 @@ shared_scripts {
 
 client_scripts {
     '@qbx_core/modules/playerdata.lua',
-    'client/main.lua'
+    'client/main.lua',
+    'client/carjack.lua',
 }
 
 server_scripts {
